% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bodytemp}
\alias{bodytemp}
\title{Human body temperature}
\format{
A data frame with 25 rows and 2 variables:
\describe{
\item{individual}{numeric. Person identifier, 1-25. Not missing.}
\item{temperature}{numeric. Body temperature in degrees Fahrenheit;
range 97.4-100.0. Not missing.}
}
}
\source{
W&S Example 11.3. Whitlock, M. & Schluter, D. (2018). Datasets for
Whitlock & Schluter's "The Analysis of Biological Data". CC BY-NC-SA 4.0.
}
\usage{
bodytemp
}
\description{
Body temperature of 25 randomly chosen healthy people. Used to teach the
one-sample t-test, notably against the conventional value of 98.6 degrees F.
}
\keyword{datasets}
