% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{flowers}
\alias{flowers}
\title{Proportion of fertile flowers}
\format{
A data frame with 30 rows and 1 variable:
\describe{
\item{prop_fertile}{numeric. Proportion of flowers that were fertile;
unitless, range 0.006-0.987. Not missing.}
}
}
\source{
Course teaching dataset, BIOL 202, UBC Okanagan.
}
\usage{
flowers
}
\description{
Proportion of flowers that were fertile in each of 30 plants. Because the
values are proportions bounded at 0 and 1, this dataset is used to teach the
arcsine square-root transformation.
}
\keyword{datasets}
