% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{haircolour}
\alias{haircolour}
\title{Student hair colour}
\format{
A data frame with 100 rows and 1 variable:
\describe{
\item{colour}{factor. Hair colour: \code{Black}, \code{Blonde}, \code{Brown}, \code{Red}.
Not missing.}
}
}
\source{
Course teaching dataset, BIOL 202, UBC Okanagan.
}
\usage{
haircolour
}
\description{
Hair colour recorded for 100 students. Used to practise frequency and
relative-frequency tables.
}
\keyword{datasets}
