% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wolf}
\alias{wolf}
\title{Wolf inbreeding and reproduction}
\format{
A data frame with 24 rows and 2 variables:
\describe{
\item{inbreed_coef}{numeric. Inbreeding coefficient; unitless, 0-0.4.
Higher values mean more inbred. Not missing.}
\item{n_pups}{numeric. Number of pups surviving their first winter;
counts, range 1-8. Not missing.}
}
}
\source{
W&S Example 16.2. Whitlock, M. & Schluter, D. (2018). Datasets for
Whitlock & Schluter's "The Analysis of Biological Data". CC BY-NC-SA 4.0.
}
\usage{
wolf
}
\description{
Inbreeding coefficient and number of pups surviving their first winter, for
24 wolf packs. Used to teach correlation.
}
\keyword{datasets}
